Billing Financial Model - Cleanup; Image Serializer, API; build folder added for staff static ,gitignore - build path bug removed

+++ src/billing/utils/fee_models/model_A.py
@@ -5,6 +5,8 @@ Fee model
     - tenant is charged service fee only for 4 weeks or less
     - home owner is charged service fee for the whole booking
 """
+from decimal import Decimal
+
 from billing.utils.fee_models.base import TenantAccountBase, HomeOwnerAccountBase, FeeModelBase, Charge
 
 
@@ -12,7 +14,7 @@ class TenantAccount(TenantAccountBase):
 
     @property
     def payable_rent(self):
-        return self.weekly_rent * min(4, self.stay_duration)
+        return Decimal(self.weekly_rent * min(4, self.stay_duration))
 
 
 class HomeOwnerAccount(HomeOwnerAccountBase):
@@ -20,7 +22,7 @@ class HomeOwnerAccount(HomeOwnerAccountBase):
     @property
     def payable_rent(self):
         """ Rent paid via Rentality to home_owner """
-        return self.weekly_rent * min(4, self.stay_duration)
+        return Decimal(self.weekly_rent * min(4, self.stay_duration))
 
     @property
     def service_fee(self):
